ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML Libraries
torch==2.8.0
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.12.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==3.17.0

# PDF Generation
reportlab==4.0.7
qrcode[pil]==7.4.2
Pillow==10.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Database
motor==3.3.2
pymongo==4.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

